requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
requests
numpy